requirements.txt
pytest
anyio
telethon
python-dotenv
